% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_katanga_population_sizes.r
\name{estimate_population_sizes}
\alias{estimate_population_sizes}
\title{Estimate population size per health zone in Katanga}
\usage{
estimate_population_sizes(under_5_multiplier = 0.169)
}
\arguments{
\item{under_5_multiplier}{factor population size is multiplied by to estimate population size under 5}
}
\value{
data frame with estimated population size ('population') and population size less than 5 years old ('l5pop') per year and per health zone
}
\description{
Estimate population size per health zone in Katanga
}
\author{
Sebastian Funk
}
