% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure1.r
\name{figure1}
\alias{figure1}
\title{Plot figure 1}
\usage{
figure1(shapefile_dir)
}
\arguments{
\item{shapefile_dir}{directory containing the shape files}
}
\value{
plot
}
\description{
Shapefiles can be obtained from
https://data.humdata.org/dataset/dr-congo-health-0. Download
RDC_Zone_de_sante.zip, unpack it and pass the directory in which it was
unpacked to this function.
}
\author{
Sebastian Funk
}
