% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure5.r
\name{figure5}
\alias{figure5}
\title{Plot figure 4}
\usage{
figure5(m)
}
\arguments{
\item{m}{a prediction model of class \code{glm}, such as returned by \code{\link{prediction_model}}}
}
\value{
plot
}
\description{
Plot figure 4
}
\author{
Sebastian Funk
}
