% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure6.r
\name{figure6}
\alias{figure6}
\title{Plot figure 6}
\usage{
figure6(posterior, posterior_no_mvc)
}
\arguments{
\item{posterior}{a \code{libbi} object containing posterior samples}

\item{posterior_no_mvc}{a \code{libbi} object containing posterior predictions without a mass-vaccination campaign}
}
\value{
plot
}
\description{
Plot figure 6
}
\author{
Sebastian Funk
}
