% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dynamic_model.r
\name{fit_dynamic_model}
\alias{fit_dynamic_model}
\title{Fit a dynamic model to the measles outbreak in Katanga}
\usage{
fit_dynamic_model(model, year = 2015, end = 5, start_threshold = 10,
  vaccine_delay = 0, nbdata, particles, under_5_multiplier = 0.169)
}
\arguments{
\item{model}{a 'bi_model'}

\item{year}{the year to which to fit}

\item{end}{how many weeks to retain after the last data point}

\item{start_threshold}{when this number is exceeded over a two-week period,
start fitting}

\item{vaccine_delay}{delay in vaccine protection; default: 0 (instantaneous protection)}

\item{nbdata}{number of data points to require; if fewer data points are
available in a health zone, the health zone will be discarded}

\item{particles}{number of particles, if given; otherwise will be determined
from the variance of the log-likelihood estimate}

\item{under_5_multiplier}{multiplier for population size to get under-5
population size; default: 0.169}
}
\value{
libbi object containing the posterior samples
}
\description{
Uses the LibBi interface provided by the rbi package
}
\author{
Sebastian Funk
}
