% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_2015_data.r
\name{gather_2015_data}
\alias{gather_2015_data}
\title{Gather 2015 data for cases, coverage and campaigns}
\usage{
gather_2015_data()
}
\value{
data frame
}
\description{
Gather 2015 data for cases, coverage and campaigns
}
\author{
Sebastian Funk
}
\keyword{internal}
