% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_quantile.r
\name{my_quantile}
\alias{my_quantile}
\title{Quantile function that works with NA}
\usage{
my_quantile(x, probs, ...)
}
\arguments{
\item{x}{numeric vector whose sample quantiles are wanted, or an
    object of a class for which a method has been defined (see also
    \sQuote{details}). \code{\link{NA}} and \code{NaN} values are not
    allowed in numeric vectors unless \code{na.rm} is \code{TRUE}.}

\item{probs}{numeric vector of probabilities with values in
    \eqn{[0,1]}.  (Values up to \samp{2e-14} outside that
    range are accepted and moved to the nearby endpoint.)}

\item{...}{further arguments passed to or from other methods.}
}
\value{
quantiles
}
\description{
Quantile function that works with NA
}
\author{
Sebastian Funk
}
\keyword{internal}
