% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_model.r
\name{prediction_model}
\alias{prediction_model}
\title{Prediction model for outbreaks in 2015}
\usage{
prediction_model(provinces = FALSE, campaigns_2015 = FALSE)
}
\arguments{
\item{campaigns_2015}{whether to include campaigns in 2015 as a variable}
}
\value{
glm object
}
\description{
Prediction model for outbreaks in 2015
}
\author{
Sebastian Funk
}
