% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_mvc.r
\name{remove_mvc}
\alias{remove_mvc}
\title{Remove mass vaccination campaigns from a LibBi model fit}
\usage{
remove_mvc(posterior, ...)
}
\arguments{
\item{posterior}{a \code{libbi} object containing posterior samples}
}
\value{
a \code{libbi} object with updated posterior predictive samples, with all mass vaccination campaigns removed
}
\description{
Remove mass vaccination campaigns from a LibBi model fit
}
\author{
Sebastian Funk
}
